wip: renamed chat to townhallchat and moved pane into components folder
