+++ src/components/Chat/Chat.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Grid } from '@material-ui/core';
 
-import { Props as ChatContentProps } from '../ChatContent';
-import { Props as ChatbarProps } from '../Chatbar';
+import { Props as ChatContentProps } from 'components/ChatContent';
+import { Props as ChatbarProps } from 'components/Chatbar';
 
 interface Props {
     children: React.ReactElement<ChatContentProps | ChatbarProps>[];
