+++ src/components/PaneContent/index.ts
