--- src/domains/Chat/ChatWrapper/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ChatWrapper';
