+++ src/domains/Townhall/TownhallChat/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TownhallChat';
