+++ src/domains/Townhall/TownhallChat/utils.ts
